test: 게시글 등록 및 댓글 관련 api 테스트 완

+++ src/main/java/com/kaboot/community/domain/board/service/impl/BoardCommandServiceImpl.java
@@ -1,25 +1,23 @@
 package com.kaboot.community.domain.board.service.impl;
 
+import com.kaboot.community.common.enums.CustomResponseStatus;
+import com.kaboot.community.common.exception.CustomException;
 import com.kaboot.community.domain.board.dto.request.CommentPostOrModifyRequest;
 import com.kaboot.community.domain.board.dto.request.LikeRequest;
 import com.kaboot.community.domain.board.dto.request.PostOrModifyRequest;
 import com.kaboot.community.domain.board.entity.Board;
 import com.kaboot.community.domain.board.entity.Comment;
 import com.kaboot.community.domain.board.entity.Likes;
-import com.kaboot.community.domain.board.mapper.BoardMapper;
 import com.kaboot.community.domain.board.mapper.CommentMapper;
 import com.kaboot.community.domain.board.mapper.LikesMapper;
+import com.kaboot.community.domain.board.repository.board.BoardRepository;
 import com.kaboot.community.domain.board.repository.comment.CommentRepository;
 import com.kaboot.community.domain.board.repository.likes.LikesRepository;
-import com.kaboot.community.domain.board.repository.board.BoardRepository;
 import com.kaboot.community.domain.board.service.BoardCommandService;
 import com.kaboot.community.domain.board.service.BoardQueryService;
-import com.kaboot.community.common.enums.CustomResponseStatus;
-import com.kaboot.community.common.exception.CustomException;
 import com.kaboot.community.domain.member.entity.Member;
 import com.kaboot.community.domain.member.service.member.MemberQueryService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,26 +36,31 @@ public class BoardCommandServiceImpl implements BoardCommandService {
 
     public void postBoard(String authUsername, PostOrModifyRequest postRequest) {
         Member validMember = getMemberByUsername(authUsername);
-        Board board = BoardMapper.toBoardFromPostRequest(postRequest, validMember);
+        Board board = postRequest.toEntity(validMember);
 
         boardRepository.save(board);
     }
 
-    @PreAuthorize("isAuthenticated() and hasAuthority('MEMBER') and authentication.principal.username == #authUsername")
     public void modifyBoard(String authUsername, PostOrModifyRequest modifyRequest, Long boardId) {
         Board validBoard = getBoardById(boardId);
+        if (!validBoard.canAccess(authUsername)) {
+            throw new CustomException(CustomResponseStatus.ACCESS_DENIED);
+        }
 
         validBoard.update(modifyRequest);
     }
 
-    @PreAuthorize("isAuthenticated() and hasAuthority('MEMBER') and authentication.principal.username == #authUsername")
     public void deleteBoard(String authUsername,  Long boardId) {
         Board validBoard = getBoardById(boardId);
+        if (!validBoard.canAccess(authUsername)) {
+            throw new CustomException(CustomResponseStatus.ACCESS_DENIED);
+        }
 
         boardRepository.delete(validBoard);
     }
 
     public void toggleLike(String authUsername, Long boardId, LikeRequest likeRequest) {
+        // Todo : 테스트 진행 필요
         Member validMember = getMemberByUsername(authUsername);
         Board validBoard = getBoardById(boardId);
 
@@ -82,16 +85,20 @@ public class BoardCommandServiceImpl implements BoardCommandService {
         commentRepository.save(CommentMapper.toEntity(validBoard, validMember, commentPostRequest));
     }
 
-    @PreAuthorize("isAuthenticated() and hasAuthority('MEMBER') and authentication.principal.username == #authUsername")
     public void modifyComment(String authUsername, Long commentId, CommentPostOrModifyRequest commentModifyRequest) {
         Comment validComment = getCommentById(commentId);
+        if (!validComment.canAccess(authUsername)) {
+            throw new CustomException(CustomResponseStatus.ACCESS_DENIED);
+        }
 
         validComment.updateComment(commentModifyRequest);
     }
 
-    @PreAuthorize("isAuthenticated() and hasAuthority('MEMBER') and authentication.principal.username == #authUsername")
     public void deleteComment(String authUsername, Long commentId) {
         Comment validComment = getCommentById(commentId);
+        if (!validComment.canAccess(authUsername)) {
+            throw new CustomException(CustomResponseStatus.ACCESS_DENIED);
+        }
 
         validComment.delete(LocalDateTime.now());
     }
